+++ Biblioteca/src/Almacen/TestLeerXML.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?><Catalogo><Book id="bk101"><Author>Garghentini, Davide</Author></Book></Catalogo>
