--- Biblioteca/src/Errores/Logging.txt.3
@@ -1,2 +1,0 @@
-oct. 23, 2020 9:15:47 A. M. reto.Escritor escribir
-WARNING: Metodo escribir fallo, Archivo No encontrado, revise la extension
