+++ Biblioteca/src/main/java/reto/Aplicacion.java
@@ -5,7 +5,6 @@ public class Aplicacion {
 	public static void main(String[] args) {
 		VentanaSeleccionArchivo iniciar = new VentanaSeleccionArchivo();
 		iniciar.iniciarVentana();
-
 	}
 
 }
